Changed start_pts to cater for an edge case found in some encoders where the start time is greater than an int

+++ src/main/java/net/bramp/ffmpeg/probe/FFmpegStream.java
@@ -44,7 +44,8 @@ public class FFmpegStream {
   public Fraction avg_frame_rate;
   public Fraction time_base;
 
-  public int start_pts;
+  // Edge case found on encoder, changed from int to double
+  public double start_pts;
   public double start_time;
 
   public long duration_ts;
